--- a/python/llm/example/CPU/QLoRA-FineTuning/alpaca-qlora/alpaca_qlora_finetuning_cpu.py
+++ b/python/llm/example/CPU/QLoRA-FineTuning/alpaca-qlora/alpaca_qlora_finetuning_cpu.py
@@ -181,7 +181,7 @@ def train(
         int4_config = BitsAndBytesConfig(
             load_in_4bit=True,
             bnb_4bit_quant_type="sym_int4",
-            bnb_4bit_compute_dtype=torch.bfloat16
+            bnb_4bit_compute_dtype=torch.float16
         )
         model = AutoModelForCausalLM.from_pretrained(base_model,
                                                      quantization_config=int4_config, )