--- a/python/llm/example/CPU/QLoRA-FineTuning/alpaca-qlora/alpaca_qlora_finetuning_cpu.py
+++ b/python/llm/example/CPU/QLoRA-FineTuning/alpaca-qlora/alpaca_qlora_finetuning_cpu.py
@@ -178,18 +178,18 @@ def train(
         )
     else:
         # Load the base model from a directory or the HF Hub to 4-bit NormalFloat format
-        int4_config = BitsAndBytesConfig(
+        bnb_config = BitsAndBytesConfig(
             load_in_4bit=True,
             bnb_4bit_use_double_quant=False,
             bnb_4bit_quant_type="int4",
             bnb_4bit_compute_dtype=torch.bfloat16
         )
         model = AutoModelForCausalLM.from_pretrained(base_model,
-                                                     quantization_config=int4_config, )
-        # also support now
+                                                     quantization_config=bnb_config, )
+        # below is also supported
         # model = AutoModelForCausalLM.from_pretrained(
         #     base_model,
-        #     load_in_low_bit="sym_int4", # not support "nf4"
+        #     load_in_low_bit="sym_int4", # nf4 not supported on cpu yet
         #     optimize_model=False,
         #     torch_dtype=torch.bfloat16,
         #     modules_to_not_convert=["lm_head"],