--- a/python/llm/example/CPU/QLoRA-FineTuning/qlora_finetuning_cpu.py
+++ b/python/llm/example/CPU/QLoRA-FineTuning/qlora_finetuning_cpu.py
@@ -49,6 +49,7 @@ if __name__ == "__main__":
 
     int4_config = BitsAndBytesConfig(
         load_in_4bit=True,
+        bnb_4bit_use_double_quant=False,
         bnb_4bit_quant_type="int4",
         bnb_4bit_compute_dtype=torch.bfloat16
     )