@rnwang04 2023-12-26
I guess this needs to rebase main branch first, as there are several updates in GPU alpaca_qlora_finetuning.py, esp we add two training_mode (lora / relora).