@hzjane 2024-01-04
On one tile of arc770, test alpaca_qlora example. The speed and memory is recorded at step 10. 
The performance and memory are basically unchanged
status | model | low_bit | Micro   batch size | Batch   size | Lora   modules | gc | speed | memory | comment
-- | -- | -- | -- | -- | -- | -- | -- | -- | --
old | Llama2-7b | int4 | 2 | 128 | full | False | 45.41s | 9526 | 
old | Llama2-7b | nf4 | 2 | 128 | full | False | 44.60s | 11717 | 
old | Llama2-7b | bf16 | 1 | 128 | full | False | 48.46s | 16287 | 
new | Llama2-7b | int4 | 2 | 128 | full | False | 46.06s | 9553 | 
new | Llama2-7b | nf4 | 2 | 128 | full | False | 44.76s | 11293 | 
new | Llama2-7b | bf16   | 1 | 128 | full | False | 48.01s | 16229 |